# api/openapi.yaml
openapi: 3.0.3
info:
  title: Animal Photos API
  description: A backend service to retrieve animal breed photos. Currently supports dogs via the Dog CEO API, extensible for other animals.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/animals/{animal}/breeds/{breed}/photo:
    get:
      summary: Get a photo for a breed
      description: Returns a random photo URL for the specified breed
      operationId: getBreedPhoto
      tags:
        - Photos
      parameters:
        - name: animal
          in: path
          required: true
          description: The animal type (e.g., dog)
          schema:
            type: string
          example: dog
        - name: breed
          in: path
          required: true
          description: The breed name
          schema:
            type: string
          example: husky
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponse'
              example:
                data:
                  url: "https://images.dog.ceo/breeds/husky/n02110185_10047.jpg"
        '404':
          description: Animal or breed not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "animal not found"
        '500':
          description: Failed to fetch photo from upstream API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    PhotoResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            url:
              type: string
              format: uri
              description: URL of the breed photo
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message